LayoutTests/accessibility/row-with-aria-role-in-native-table-expected.txt

Item 1
This tests if a native tr element uses the 'row' role, but forgets to add role='grid' to the native table element, the table still exposes the right numbers of rows/cells.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table.rowCount is 1
PASS table.columnCount is 1
PASS successfullyParsed is true

TEST COMPLETE
